Better organize sharepoint methods
