+++ server/src/document/document.module.ts
@@ -3,11 +3,13 @@ import { DocumentController } from './document.controller';
 import { ConfigModule } from '../config/config.module';
 import { CrmModule } from '../crm/crm.module';
 import { DocumentService } from './document.service';
+import { SharepointModule } from '../sharepoint/sharepoint.module';
 
 @Module({
   imports: [
     ConfigModule,
     CrmModule,
+    SharepointModule,
   ],
   providers: [
     DocumentService,
