+++ server/src/document/document.service.ts
@@ -4,6 +4,12 @@ import { CrmService } from '../crm/crm.service';
 import { ADAL } from '../_utils/adal';
 import * as Request from 'request';
 
+
+// This service currently provides utilities to write files and folders
+// to Sharepoint via CRM. It uses fetchXML api and an undocumented
+// /Upload endpoint on the agency's custom Dynamics CRM.
+// Isolating this service allows us to eventually deprecate it
+// in favor of the Sharepoint service.
 @Injectable()
 export class DocumentService {
   crmUrlPath = '';
@@ -278,35 +284,5 @@ async getParentSiteLocation() {
         }
       });
     });
-  }
-
-  // We retrieve documents from the Sharepoint folder (`folderIdentifier` in the
-  // code below) that holds both documents from past revisions and the current
-  // revision. CRM automatically carries over documents from past revisions into
-  // this folder, and we deliberately upload documents for the latest/current
-  // revision into this folder.
-  async findPackageSharepointDocuments(packageName, id: string) {
-    try {
-      const folderIdentifier = `${packageName}_${id.toUpperCase().replace(/-/g, '')}`;
-
-      const { value: documents } = await this.crmService.getSharepointFolderFiles(`dcp_package/${folderIdentifier}`);
-
-      return documents;
-    } catch (e) {
-      // Relay errors from crmService 
-      if (e instanceof HttpException) {
-        throw e;
-      } else {
-        throw new HttpException({
-          code: 'SHAREPOINT_FOLDER_ERROR',
-          title: 'Bad Sharepoint folder lookup',
-          detail: `An error occured while constructing and looking up folder for package. Perhaps the package name or id is wrong.`,
-          meta: {
-            packageName: packageName,
-            packageId: id,
-          }
-        }, HttpStatus.INTERNAL_SERVER_ERROR);
-      }
-    }
   }
 }
